Move mobile overlay into app
So when programs or channels reload, it doesn't show again
